+++ R/matrix.R
@@ -59,14 +59,17 @@ PDsolve <- function(M,force=FALSE,pseudo=FALSE)
   DIM <- dim(M)
   if(is.null(DIM)) { DIM <- 1 }
 
-  if(DIM[1]==1) { return(1/M) }
-  if(DIM[1]==2)
+  if(!force && !pseudo)
   {
-    DET <- M[1,1]*M[2,2]-M[1,2]*M[2,1]
-    SWP <- M[1,1] ; M[1,1] <- M[2,2] ; M[2,2] <- SWP
-    M[1,2] <- -M[1,2]
-    M[2,1] <- -M[2,1]
-    return( M/DET )
+    if(DIM[1]==1) { return(1/M) }
+    if(DIM[1]==2)
+    {
+      DET <- M[1,1]*M[2,2]-M[1,2]*M[2,1]
+      SWP <- M[1,1] ; M[1,1] <- M[2,2] ; M[2,2] <- SWP
+      M[1,2] <- -M[1,2]
+      M[2,1] <- -M[2,1]
+      return( M/DET )
+    }
   }
 
   TOL <- DIM[1]*.Machine$double.eps
@@ -85,7 +88,7 @@ PDsolve <- function(M,force=FALSE,pseudo=FALSE)
   # try ordinary inverse
   M.try <- try(qr.solve(M,tol=TOL))
   # fall back on decomposition
-  if(class(M.try) == "matrix")
+  if( class(M.try) == "matrix")
   { M <- M.try }
   else
   { M <- PDfunc(M,func=function(m){1/m},force=force,pseudo=pseudo) }